scripts/requirements.txt
supabase>=1.0.3
python-dotenv>=1.0.0
openai>=1.0.0
reportlab>=4.0.0
matplotlib>=3.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
PyYAML>=6.0
